Complete Ch03

+++ chapters/ch03-scalars/examples/fixme_02.rs
@@ -12,11 +12,11 @@ fn main() {
     // A file-size counter: u32 is too small for the total, so it's accumulated in a u64.
     let file_a: u32 = 3_000_000_000;
     let file_b: u32 = 2_000_000_000;
-    let total: u64 = file_a + file_b;
+    let total: u64 = file_a as u64 + file_b as u64;
     println!("total bytes: {total}");
 
     let sum: i32 = 10;
     let count: i32 = 4;
-    let average: f64 = sum / count;
+    let average: f64 = sum as f64 / count as f64;
     println!("average: {average}");
 }
